Remove RecoverableDeciderExceptionInterface
