+++ src/Exception/MachineProvider/MissingRemoteMachineException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Exception\MachineProvider;
+
+use App\Exception\MachineProvider\AbstractNotFoundRemoteMachineException as Base;
+use App\Exception\UnrecoverableExceptionInterface;
+
+class MissingRemoteMachineException extends Base implements UnrecoverableExceptionInterface
+{
+}
