+++ src/Exception/MachineProvider/NotFoundRemoteMachineExceptionInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Exception\MachineProvider;
+
+use App\Enum\MachineProvider;
+
+interface NotFoundRemoteMachineExceptionInterface extends ExceptionInterface
+{
+    public function getProvider(): MachineProvider;
+}
