--- src/Exception/RecoverableDeciderExceptionInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Exception;
-
-interface RecoverableDeciderExceptionInterface
-{
-    public function isRecoverable(): bool;
-}
