+++ src/MessageHandler/CreateMachineHandler.php
@@ -6,8 +6,6 @@ namespace App\MessageHandler;
 
 use App\Entity\Machine;
 use App\Enum\MachineState;
-use App\Exception\MachineActionFailedException;
-use App\Exception\RecoverableDeciderExceptionInterface;
 use App\Exception\UnrecoverableExceptionInterface;
 use App\Message\CreateMachine;
 use App\Repository\MachineRepository;
@@ -45,24 +43,8 @@ class CreateMachineHandler
             $remoteMachine = $this->machineManager->create($machine);
             $this->machineUpdater->updateFromRemoteMachine($machine, $remoteMachine);
             $this->machineRequestDispatcher->dispatchCollection($message->getOnSuccessCollection());
-        } catch (MachineActionFailedException $exception) {
-            throw new UnrecoverableMessageHandlingException(
-                $exception->getMessage(),
-                $exception->getCode(),
-                $exception
-            );
-        } catch (\Throwable $exception) {
-            if (
-                $exception instanceof UnrecoverableExceptionInterface
-                || $exception instanceof RecoverableDeciderExceptionInterface && false === $exception->isRecoverable()
-            ) {
-                $code = $exception->getCode();
-                $code = is_int($code) ? $code : 0;
-
-                throw new UnrecoverableMessageHandlingException($exception->getMessage(), $code, $exception);
-            }
-
-            throw $exception;
+        } catch (UnrecoverableExceptionInterface $e) {
+            throw new UnrecoverableMessageHandlingException($e->getMessage(), $e->getCode(), $e);
         }
     }
 }
