+++ src/MessageHandler/GetMachineHandler.php
@@ -5,8 +5,6 @@ declare(strict_types=1);
 namespace App\MessageHandler;
 
 use App\Entity\Machine;
-use App\Exception\MachineActionFailedException;
-use App\Exception\RecoverableDeciderExceptionInterface;
 use App\Exception\UnrecoverableExceptionInterface;
 use App\Message\GetMachine;
 use App\Repository\MachineRepository;
@@ -41,24 +39,8 @@ class GetMachineHandler
             $remoteMachine = $this->machineManager->get($machine);
             $this->machineUpdater->updateFromRemoteMachine($machine, $remoteMachine);
             $this->machineRequestDispatcher->dispatchCollection($message->getOnSuccessCollection());
-        } catch (MachineActionFailedException $exception) {
-            throw new UnrecoverableMessageHandlingException(
-                $exception->getMessage(),
-                $exception->getCode(),
-                $exception
-            );
-        } catch (\Throwable $exception) {
-            if (
-                $exception instanceof UnrecoverableExceptionInterface
-                || $exception instanceof RecoverableDeciderExceptionInterface && false === $exception->isRecoverable()
-            ) {
-                $code = $exception->getCode();
-                $code = is_int($code) ? $code : 0;
-
-                throw new UnrecoverableMessageHandlingException($exception->getMessage(), $code, $exception);
-            }
-
-            throw $exception;
+        } catch (UnrecoverableExceptionInterface $e) {
+            throw new UnrecoverableMessageHandlingException($e->getMessage(), $e->getCode(), $e);
         }
     }
 }
